requirements.txt
Flask
paho-mqtt
wakeonlan
requests
